+++ src/systems/TextAndGLDebug.h
@@ -9,10 +9,10 @@ namespace Core
 		namespace TextAndGLDebug
 		{
 			void Init();
-			void Render(int _w, int _h);
+			void Render(float _w, float _h);
 			void Cleanup();
 
-			bool RenderText(int32 _fontI, fVec2 _tlPos, char const* _text, float _size = 16.0f, uint32 _col = COL_WHITE);
+			bool RenderText(uint32 _fontI, fVec2 _tlPos, char const* _text, float _size = 16.0f, uint32 _col = COL_WHITE);
 		}
 	}
 }
